move base_convolution code to a separate cpp file
